Adds `build-essential` debian package
I have a use case where I'm building a rust library for use by android. Rust needs a c compiler on the path and without it I can't use jitpack, which I really want to because it seems like exactly what I need.

If this not something you want to support, that's all good. Just thought it might be an easy change.